test fixes and change analyzer settings for uwp
