+++ tests/TestData/SmokeTests.WinUI3UWPBaseline/App.xaml.cs
@@ -11,14 +11,14 @@ using Windows.ApplicationModel;
 using Windows.ApplicationModel.Activation;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using Microsoft.UI.Xaml.Data;
-using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Media.Animation;
-using Microsoft.UI.Xaml.Navigation;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
+using Windows.UI.Xaml.Controls.Primitives;
+using Windows.UI.Xaml.Data;
+using Windows.UI.Xaml.Input;
+using Windows.UI.Xaml.Media;
+using Windows.UI.Xaml.Media.Animation;
+using Windows.UI.Xaml.Navigation;
 
 namespace OldCsProj
 {
@@ -42,7 +42,7 @@ namespace OldCsProj
         /// will be used such as when the application is launched to open a specific file.
         /// </summary>
         /// <param name="e">Details about the launch request and process.</param>
-        protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs e)
+        protected override void OnLaunched(LaunchActivatedEventArgs e)
         {
             Frame rootFrame = Window.Current.Content as Frame;
             TestAnalyzers();
@@ -55,7 +55,7 @@ namespace OldCsProj
 
                 rootFrame.NavigationFailed += OnNavigationFailed;
 
-                if (e.UWPLaunchActivatedEventArgs.PreviousExecutionState == ApplicationExecutionState.Terminated)
+                if (e.PreviousExecutionState == ApplicationExecutionState.Terminated)
                 {
                     //TODO: Load state from previously suspended application
                 }
@@ -64,21 +64,21 @@ namespace OldCsProj
                 Window.Current.Content = rootFrame;
             }
 
-            if (e.UWPLaunchActivatedEventArgs.PrelaunchActivated == false)
+            if (e.PrelaunchActivated == false)
             {
                 if (rootFrame.Content == null)
                 {
                     // When the navigation stack isn't restored navigate to the first page,
                     // configuring the new page by passing required information as a navigation
                     // parameter
-                    rootFrame.Navigate(typeof(MainPage), e.UWPLaunchActivatedEventArgs.Arguments);
+                    rootFrame.Navigate(typeof(MainPage), e.Arguments);
                 }
                 // Ensure the current window is active
                 Window.Current.Activate();
             }
         }
 
-        private class oldINotifyPropertyChanged : Microsoft.UI.Xaml.Data.INotifyPropertyChanged
+        private class oldINotifyPropertyChanged : INotifyPropertyChanged
         {
             event PropertyChangedEventHandler INotifyPropertyChanged.PropertyChanged
             {
@@ -93,7 +93,7 @@ namespace OldCsProj
                 }
             }
         }
-        private class oldICommand : Microsoft.UI.Xaml.Input.ICommand
+        private class oldICommand : ICommand
         {
             event EventHandler ICommand.CanExecuteChanged
             {
@@ -121,17 +121,17 @@ namespace OldCsProj
 
         void TestAnalyzers()
         {
-            PropertyChangedEventArgs changeArgs = new Microsoft.UI.Xaml.Data.PropertyChangedEventArgs("tstProp");
+            PropertyChangedEventArgs changeArgs = new PropertyChangedEventArgs("tstProp");
             ObservableCollection<string> c = new ObservableCollection<string>(); //Cannot run in try-convert
-            var cornerRadius1 = Microsoft.UI.Xaml.CornerRadiusHelper.FromUniformRadius(2);
-            var cornerRadius2 = Microsoft.UI.Xaml.CornerRadiusHelper.FromRadii(1,2,3,4);
-            var duration = Microsoft.UI.Xaml.DurationHelper.FromTimeSpan(new TimeSpan(33));
-            var gridLength1 = Microsoft.UI.Xaml.GridLengthHelper.FromPixels(12);
-            var gridLength2 = Microsoft.UI.Xaml.GridLengthHelper.FromValueAndType(1, GridUnitType.Pixel);
-            var thickness1 = Microsoft.UI.Xaml.ThicknessHelper.FromUniformLength(2);
-            var thickness2 = Microsoft.UI.Xaml.ThicknessHelper.FromLengths(2, 3, 4, 5);
-            var generatorPosition = Microsoft.UI.Xaml.Controls.Primitives.GeneratorPositionHelper.FromIndexAndOffset(23, 123);
-            var matrix = Microsoft.UI.Xaml.Media.MatrixHelper.FromElements(6, 5, 4, 3, 2, 1);
+            var cornerRadius1 = new CornerRadius(2);
+            var cornerRadius2 = new CornerRadius(1,2,3,4);
+            var duration = new Duration(new TimeSpan(33));
+            var gridLength1 = new GridLength(12);
+            var gridLength2 = new GridLength(1, GridUnitType.Pixel);
+            var thickness1 = new Thickness(2);
+            var thickness2 = new Thickness(2, 3, 4, 5);
+            var generatorPosition = new GeneratorPosition(23, 123);
+            var matrix = new Matrix(6, 5, 4, 3, 2, 1);
             var repeatBehavior1 = new RepeatBehavior(12);
             var repeatBehavior2 = new RepeatBehavior(new TimeSpan(10));
         }
